+++ plugins/templates/policies/frontends/policies.ts
@@ -6,27 +6,29 @@
  */
 
 import { IPolicy } from '@models/policy/types';
-import { Types } from 'mongoose';
+import { objectID } from '@plugins/misc/uid';
 
-const objectID = (num: string) => Types.ObjectId(`pol-fend@${num}`);
-
+const CODE = 'ftend';
 export const frontend_add: Readonly<IPolicy> = {
-  _id: objectID('001'),
+  _id: objectID('p'),
+  code: `${CODE}add`,
   name: 'Add Frontends',
   message: 'Enable Adding of Frotnends from Database',
   global_flag: true,
 };
 
 export const frontend_edit: Readonly<IPolicy> = {
-  _id: objectID('002'),
-  name: 'Edit Credentials',
+  _id: objectID('p'),
+  code: `${CODE}edit`,
+  name: 'Edit Frontends',
   message: 'Enable Editing of Frontend Details from Database',
   global_flag: true,
 };
 
 export const frontend_rm: Readonly<IPolicy> = {
-  _id: objectID('003'),
-  name: 'Remove Credentials',
+  _id: objectID('p'),
+  code: `${CODE}rm`,
+  name: 'Remove Frontends',
   message: 'Enable Removal of Frontends from Database',
   global_flag: true,
 };