+++ plugins/templates/policies/roles/policies.ts
@@ -6,26 +6,29 @@
  */
 
 import { IPolicy } from '@models/policy/types';
-import { Types } from 'mongoose';
+import { objectID } from '@plugins/misc/uid';
 
-const objectID = (num: string) => Types.ObjectId(`pol-role@${num}`);
+const CODE = 'role';
 
 export const roles_add: Readonly<IPolicy> = {
-  _id: objectID('001'),
+  _id: objectID('p'),
+  code: `${CODE}add`,
   name: 'Add Sub Roles',
   message: 'Enable Creating of Sub Roles',
   global_flag: true,
 };
 
 export const roles_edit: Readonly<IPolicy> = {
-  _id: objectID('002'),
+  _id: objectID('p'),
+  code: `${CODE}edit`,
   name: 'Edit Sub Roles',
   message: 'Enable Editing of Settings related to Sub Roles',
   global_flag: true,
 };
 
 export const roles_rm: Readonly<IPolicy> = {
-  _id: objectID('003'),
+  _id: objectID('p'),
+  code: `${CODE}rm`,
   name: 'Remove Sub Roles',
   message: 'Enable Removal of Sub Roles',
   global_flag: true,