+++ plugins/templates/policies/scopes/policies.ts
@@ -6,26 +6,29 @@
  */
 
 import { IPolicy } from '@models/policy/types';
-import { Types } from 'mongoose';
+import { objectID } from '@plugins/misc/uid';
 
-const objectID = (num: string) => Types.ObjectId(`pol-scop@${num}`);
+const CODE = 'scp';
 
 export const scope_add: Readonly<IPolicy> = {
-  _id: objectID('001'),
+  _id: objectID('p'),
+  code: `${CODE}add`,
   name: 'Add Scopes',
   message: 'Enable Adding of Scopes to Database',
   global_flag: true,
 };
 
 export const scope_edit: Readonly<IPolicy> = {
-  _id: objectID('002'),
+  _id: objectID('p'),
+  code: `${CODE}edit`,
   name: 'Edit Scopes',
   message: 'Enable Editing of Scopes in Database',
   global_flag: true,
 };
 
 export const scope_rm: Readonly<IPolicy> = {
-  _id: objectID('003'),
+  _id: objectID('p'),
+  code: `${CODE}rm`,
   name: 'Remove Scopes',
   message: 'Enable Removal of Scopes from Database',
   global_flag: true,