+++ plugins/templates/policies/user/modify/policies.ts
@@ -6,33 +6,37 @@
  */
 
 import type { IPolicy } from '@models/policy/types';
-import { Types } from 'mongoose';
+import { objectID } from '@plugins/misc/uid';
 
-const objectID = (num: string) => Types.ObjectId(`p-user-m@${num}`);
+const CODE = 'usr';
 
 export const viewer_policies: Readonly<IPolicy> = {
-  _id: objectID('001'),
+  _id: objectID('p'),
+  code: `${CODE}polvwr`,
   name: 'Edit Policies of Viewer',
   message: 'Enable Granting/Removing Policies to Viewers and its Delegates',
   global_flag: true,
 };
 
 export const viewer_scope: Readonly<IPolicy> = {
-  _id: objectID('002'),
+  _id: objectID('p'),
+  code: `${CODE}scpvwr`,
   name: 'Edit Scopes of Viewers',
   message: 'Enable Granting/Removing of Scopes to Viewers and its Delegates',
   global_flag: true,
 };
 
 export const viewer_restrict: Readonly<IPolicy> = {
-  _id: objectID('003'),
+  _id: objectID('p'),
+  code: `${CODE}restvwr`,
   name: 'Restrict Viewers',
   message: 'Restrict Users and its Delegates',
   global_flag: true,
 };
 
 export const content_mgr_policies: Readonly<IPolicy> = {
-  _id: objectID('004'),
+  _id: objectID('p'),
+  code: `${CODE}polcntmgr`,
   name: 'Edit Policies of Content Manager',
   message:
     'Enable Granting/Removing Policies to Content Managers and its Delegates',
@@ -40,7 +44,8 @@ export const content_mgr_policies: Readonly<IPolicy> = {
 };
 
 export const content_mgr_scope: Readonly<IPolicy> = {
-  _id: objectID('005'),
+  _id: objectID('p'),
+  code: `${CODE}scpcntmgr`,
   name: 'Edit Scopes of Content Managers',
   message:
     'Enable Granting/Removing of Scopes to Content Managers and its Delegates',
@@ -48,49 +53,56 @@ export const content_mgr_scope: Readonly<IPolicy> = {
 };
 
 export const content_mgr_restrict: Readonly<IPolicy> = {
-  _id: objectID('006'),
+  _id: objectID('p'),
+  code: `${CODE}restcntmgr`,
   name: 'Restrict Content Managers',
   message: 'Restrict Content Managers and its Delegates',
   global_flag: true,
 };
 
 export const mods_policies: Readonly<IPolicy> = {
-  _id: objectID('007'),
+  _id: objectID('p'),
+  code: `${CODE}polmdtr`,
   name: 'Edit Policies of Moderators',
   message: 'Enable Granting/Removing Policies to Moderators and its Delegates',
   global_flag: true,
 };
 
 export const mods_scope: Readonly<IPolicy> = {
-  _id: objectID('008'),
+  _id: objectID('p'),
+  code: `${CODE}scpmdtr`,
   name: 'Edit Scopes of Moderators',
   message: 'Enable Granting/Removing of Scopes to Moderators and its Delegates',
   global_flag: true,
 };
 
 export const mods_restrict: Readonly<IPolicy> = {
-  _id: objectID('009'),
+  _id: objectID('p'),
+  code: `${CODE}restmdtr`,
   name: 'Restrict Moderators',
   message: 'Restrict Moderators and its Delegates',
   global_flag: true,
 };
 
 export const mgr_policies: Readonly<IPolicy> = {
-  _id: objectID('010'),
+  _id: objectID('p'),
+  code: `${CODE}polmgr`,
   name: 'Edit Policies of Managers',
   message: 'Enable Granting/Removing Policies to Managers and its Delegates',
   global_flag: true,
 };
 
 export const mgr_scope: Readonly<IPolicy> = {
-  _id: objectID('011'),
+  _id: objectID('p'),
+  code: `${CODE}scpmgr`,
   name: 'Edit Scopes of Managers',
   message: 'Enable Granting/Removing of Scopes to Managers and its Delegates',
   global_flag: true,
 };
 
 export const mgr_restrict: Readonly<IPolicy> = {
-  _id: objectID('012'),
+  _id: objectID('p'),
+  code: `${CODE}restmgr`,
   name: 'Restrict Managers',
   message: 'Restrict Managers and its Delegates',
   global_flag: true,